Set python-version to 3.10 in Windows CI job. (#11175)
This fixes a build failure of the Python bindings on Windows. We were
testing Python 3.10.8, but now we are testing Python 3.11.0.
* Old 3.10.8 success:
https://github.com/iree-org/iree/actions/runs/3452338984/jobs/5762178428
* Old 3.11.0 failure:
https://github.com/iree-org/iree/actions/runs/3461697554/jobs/5779634638

This pins the Python version used on the CI to 3.10. We can also fix the
break by updating pybind11 to >= 2.10.1 (which will require rebuilding
our Docker images, see https://github.com/iree-org/iree/pull/11155).

Successful run:
https://github.com/iree-org/iree/actions/runs/3473912159/jobs/5806499215
(except for 1 new test failure)