Adopt the new memref lowering process
Subviews and other view like operations will soon require to be
expanded via expand-strided-metadata before being able to be converted
to the LLVM dialect (convert-memref-to-llvm).

This change is a little bit ahead of the LLVM open source curve because
the expansion is not yet required, but it doesn't hurt to do right now.

By adopting this change early we avoid holding the LLVM integration when
this requirements will be enforced (i.e., when
https://reviews.llvm.org/D136377 lands in LLVM.)