Deprecate `IREE_ENABLE_CCACHE` option (#11283)
This setting uses `RULE_LAUNCH_COMPILE` which is discouraged:
https://cmake.org/cmake/help/latest/prop_gbl/RULE_LAUNCH_COMPILE.html.
It's also unnecessary. As of CMake 3.4, the
`CMAKE_<LANG>_COMPILER_LAUNCHER` can be used to set the compiler
launcher per language. They can be set as a configuration option or via
an environment variable. There's really no reason for us to have a
project-specific configuration for this.

While I was here, I made it an error if `IREE_ENABLE_CCACHE` is set and
`ccache` is not found. This is an easy case where we can tell people to
stop using it :-)

See some discussion in https://github.com/iree-org/iree/pull/3689