Consistently query `ExecutableTargetAttr` (#11291)
#11057 added a `IREE::HAL::ExecutableTargetAttr::lookup` function which
should be the preferred way to get a `ExecutableTargetAttr`, since it
honors the overrides made possible by that PR.

It's also neatly shrinking code at call sites, and better conveys
intent: before, we were querying a `ExecutableVariantOp` only to get at
the `ExecutableTargetAttr`, now we query directly the latter.

The Utils are updated to all take a `ExecutableTargetAttr`, and since
they were already implementing logic to handle optionally present
fields, this further simplifies logic at call sites.

Some Utils implementation functions, which were internal, are now
exposed in `Utils.h`. Particularly `getConfigStringAttr`. There seemed
to be no point having various place in the code reimplement that
boilerplate themselves (with opportunities to mishandle optionals). A
new `getConfigIntegerAttr` is added.

This is preparation for determining tile sizes based on target info in
`MaterializeEncoding`, which following #11290 we will be able to do
based on the `ExecutableTargetAttr` which we query there. It would be
inconvenient, at least for testing purposes, to have to deal only with
`ExecutableVariantOp` there, so the new `MaterializeEncoding` code in
#11290 is using `IREE::HAL::ExecutableTargetAttr::lookup` and needs to
access helpers taking a `ExecutableTargetAttr`, which was the original
motivation for the `Utils.*` changes here.