EncodingInfo may depend on the shape, not just on the element type (#11300)
In #11290 I assumed that the element type was the only part of the
shaped type that the EncodingInfo would need to depend on, but I was
forgetting that in certain narrow-shape cases, the narrow dimension
values are actually relevant. The idea is that we can avoid padding a
very narrow matrix to the kernel tile size, by adjusting the kernel tile
size accordingly. The prime example is in matrix-times-vector products:
we don't want to turn the vector into a matrix of width say 8 just
because that's the width of our matrix-times-matrix kernel. The existing
matmul-to-mmt4d pass has code for this, generating narrow kernel tile
shapes, here, Compare `m0k0n0`, `m0k0n0ForMatVec`,
`m0k0n0ForWhenRhsHas2Columns`:
https://github.com/iree-org/iree/blob/d0316e29dea6f0c720ada9e7e5df30348fb4449f/compiler/src/iree/compiler/Dialect/Flow/Transforms/ConvertLinalgMatmulToMmt4D.cpp#L322-L391
.