Encode the matmul type triple in `TensorEncoding`  (#11355)
`TensorEncoding` is used to determine tile sizes in
`MaterializeEncoding`. These depend on which SIMD instructions we want
to use to perform a matmul, which in turn depends on the (LHS, RHS,
RESULT) element type triple of the matmul --- not just one tensor's
element type in isolation.

While updating `SetEncoding`, I found some opportunity to tidy up. The
pattern was performing some rewrites (padding) before it was done
potentially returning failure, so I have reordered that. And the local
variables are renamed to be more explanatory.