Get TF:XLA working in docker (#11483)
XLA:GPU and XLA:CPU baselines are being generated manually on a VM
instance with similar hardware to the runners. This PR adds CUDA and
CuDNN packages to the SHARK dockerfile in order for Tensorflow to run
with XLA in the Docker container, removing the need to run benchmarks
manually to get XLA baselines.