Preserve `encoding` attr on `PackOp` for `TileAndDistributeToWorkgroups` (#11603)
During `TileAndDistributeToWorkgroups`, in the case of dynamic tile
sizes, the `LowerDispatchWorkgroupCountFromSetEncodingOp` pattern needs
to call `materializeEncodingValueFn`, passing it a tensor type,
expecting that tensor type to have a `TensorEncodingAttr`. The problem
is that `MaterializeEncodingPass` has already run, rewriting the
`SetEncoding` op and its result tensor, which used to hold the
`TensorEncodingAttr`, into a `PackOp` op, whose new result tensor does
not anymore have a `TensorEncodingAttr`.

As a work-around for that, we make `MaterializeEncoding` preserve the
`TensorEncodingAttr` as an attribute on the `PackOp` itself, so
`TileAndDistributeToWorkgroups` can read it and reconstruct a
`tensorTypeWithEncoding`, so
`LowerDispatchWorkgroupCountFromSetEncodingOp` can call
`materializeEncodingValueFn`.