Post to Discord when CI fails (#11672)
This turns out to be a million times easier than sending emails. I don't
love it, but it's a whole lot better than the current situation.

The result currently looks like this (the bot avatar and name as well as
the channel it posts to can be configured on the Discord side):


![image](https://user-images.githubusercontent.com/5732088/209882576-96c1e9d7-91ce-4a15-84b3-11ebfceca81e.png)


Another option would be creating an issue on each failure. I like that
a lot better for tracking purposes, but I don't think it has the same
visibility and there's still no way to create rotations on GitHub, so
we couldn't auto-assign the issue easily. It would also be quite noisy.
We could always post to the same issue, but then the issue isn't that
useful for tracking.

My intent is to start this out in the #github channel and then graduate
it to the #builds channel if everything seems to be working. For now I
also went with an action that does all the Discord message formatting
for you. If we wanted more customization, we could switch to a
different one (there are quite a few). There's someone in the issues
for this action complaining that the release is unauditable:
https://github.com/sarisia/actions-status-discord/issues/364, but the
author shows how you can get the audit trail and I'm not too worried.

I guess if this is considered adequate then:
Fixes https://github.com/iree-org/iree/issues/9857

Tested: turned on runs on pushes to another branch and pushed to it.