Rework null-terminator handling in ireeCompilerSourceWrapBuffer. (#11844)
What was there was not able to take a known-binary buffer that we do not
know whether it is null terminated or not (nor do we care). This is a
more explicit API which at least has the benefit or letting us represent
known-null-terminated buffers or known-binary buffers without having
LLVM try to probe one past what may be owned memory.

This is a breaking API change, but we are not yet stable so just making
it vs versioning the entrypoint.