Replace constant `outs` operands of Linalg ops as well. (#11837)
This change replaces the `outs` operands of Linalg ops that are splat constants with `fill` of that value. Having constants operands for `outs` causes bufferization issues downstream (which are harder to fix, though in theory whether the outs is a constant or not it shouldnt matter. The general case fix looks very similar to what happens here anyway).

Fixes #11830