Simplify the build system following the removal of old `matmul-to-mmt4d` pass (#11752)
This is part 7/9 of #11755: Finish the transition to
`--iree-flow-enable-data-tiling`.

This pass used to take target information in a custom format as pass
options. Now that it's gone, the build system logic that was generating
these pass options can go. The main complication was the
`#pass_options_variant#` placeholder that the build-system was rewriting
with specific options for each target CPU features variant. That goes
away, along with some debt that existed only to support that.