Remove `CustomKernelsTargetInfo` (#11754)
This is part 9/9 of #11755: Finish the transition to
`--iree-flow-enable-data-tiling`.

This helper dated back to early mmt4d days when we didn't have a
standardized way to gather target information, and we needed it be able
to get that information from either a `HAL::ExecutableVariant` or custom
pass options that we had to parse. The switch to
`--iree-flow-enable-data-tiling` and other cleanup steps such as #11753
leaves that no longer used at all so we can drop it now.