Discard root tasks before leaf tasks. (#11853)
Fixes #11843. The assertion there is saying that we should not be
discarding a task before its dependencies. From the comments at
https://github.com/iree-org/iree/blob/e61c192ad1ac4d1106fcad1d8795d2d53a189605/runtime/src/iree/hal/drivers/local_task/task_command_buffer.c#L51-L61
I understand that root tasks may be dependencies of leaf tasks, so it
makes sense to order the `iree_task_list_discard` calls accordingly?
This fixes the assertion observed in #11843, anyway.